mstore repo
make memory store repository for logs via cmap

+++ internal/features/logs/repository/repository.go
@@ -4,7 +4,7 @@ import LogEntry "userLoggingProject/internal/features/logs/entity"
 
 type LogsRepository interface {
 	SaveAll(userId string, entries []LogEntry.LogEntry) error
-	Save(userId string, entry *LogEntry.LogEntry) (int64, error)
+	Save(userId string, entry *LogEntry.LogEntry) (int, error)
 	LoadAll(userId string) ([]LogEntry.LogEntry, error)
 	Load(userId string, count int) ([]LogEntry.LogEntry, error)
 	RemoveAll(userId string) error
